+++ R/spatialSPsurv.R
@@ -119,7 +119,7 @@ spatialSPsurv <- function(duration,
 #' @title summary.spatialSPsurv
 #' @description Returns a summary of a exchangeSPsurv object via \code{\link[coda]{summary.mcmc}}.
 #' @param object an object of class \code{spatialSPsurv}, the output of \code{\link{spatialSPsurv}}.
-#' @param parameter one of three parameters of the \code{\link{spatialSPsurv}} output. Indicate either "betas," "gammas," or "lambda."
+#' @param parameter one of five parameters of the \code{\link{spatialSPsurv}} output. Indicate either "betas," "gammas," "rho", "lambda" or "delta".
 #' @param ... additional parameter.
 #' @return list. Empirical mean, standard deviation and quantiles for each variable.
 #' @rdname spatialSPsurv
@@ -127,12 +127,8 @@ spatialSPsurv <- function(duration,
 #'
 #'
 
-summary.spatialSPsurv <- function(object, parameter = c("betas", "gammas", "lambda"), ...){
-
-    if (parameter == "betas")  sum <- summary(mcmc(object$betas),  ...)
-    if (parameter == "gammas") sum <- summary(mcmc(object$gammas), ...)
-    if (parameter == "lambda") sum <- summary(mcmc(object$lambda), ...)
-    sum
+summary.spatialSPsurv <- function(object, parameter = character(), ...){
+    summary(coda::mcmc(object[[parameter]]), ...)
 }
 
 
@@ -171,19 +167,16 @@ print.spatialSPsurv <- function(x, ...){
 #' @title plot.spatialSPsurv
 #' @description Returns a plot of a spatialSPsurv object via \code{\link[coda]{plot.mcmc}}.
 #' @param x an object of class \code{spatialSPsurv}, the output of \code{\link{spatialSPsurv}}.
+#' @param parameter one of five parameters of the \code{\link{spatialSPsurv}} output. Indicate either "betas," "gammas," "rho", "lambda" or "delta".
 #' @param ... additional parameter.
 #' @return list. Empirical mean, standard deviation and quantiles for each variable.
 #' @rdname spatialSPsurv
 #' @export
 #'
 
-plot.spatialSPsurv <- function(x, ...){
-    plot((coda::mcmc(x$betas)), ...)
+plot.spatialSPsurv <- function(x,  parameter = character(), ...){
+    plot((coda::mcmc(x[[parameter]])), ...)
 }
-
-
-
-