+++ NEWS.md
@@ -2,6 +2,7 @@
 # BayesSPsurv 0.1.2
 
 * threshold argument added to `plot_Moran.I()` and `plot_JointCount()`
+* parameter argument added to generic function `plot()`
 
 # BayesSPsurv 0.1.1